Add sudo support for extra needs
If you set the SUDOERS_CONFIG_FOLDER environ. variable to a folder
containing sudo rights definition files, those will be copied on startup
to /etc/sudoers.d/ with the proper permissions.

For instance, this should allow to give someone
extra rights like handling data for www-data. Example of a (very permissive)
access can be
   jpommier ALL=(www-data) NOPASSWD: ALL
The NOPASSWD directive seems to be necessary, I'm not sure I'm properly
retrieving the passwords from the LDAP.

I also didn't manage to assign groups from the LDAP roles, which means
that you have to assign the rules directly to a username, which is
not ideal.